Autopackage fast build mode

+++ autoPackage.py
@@ -4,6 +4,12 @@ import logging
 from pathlib import Path
 from rich.logging import RichHandler
 from rich.progress import track
+import argparse
+
+parser = argparse.ArgumentParser()
+parser.add_argument("--debug", help="Debug Build -> The build is not clean but faster", action="store_true", default=False)
+
+args = parser.parse_args()
 
 
 FORMAT = "%(message)s"
@@ -15,13 +21,15 @@ log = logging.getLogger()
 
 log.info(f"Start Auto packaging")
 
-if os.path.exists(Path("build")):
+if os.path.exists(Path("build")) and not args.debug:
     log.info(f"Build Directory already exists -> Deleting it")
     try:
         shutil.rmtree(Path("build"))
     except Exception:
         log.exception(f"Error while removing build directory")
         exit(1)
+else:
+    log.info(f"Not deleting build directory because of debug build mode")
     
 
 # Execute setup.py
@@ -43,7 +51,8 @@ try:
     ]
     
     for i in track(files, description="Copying the packagingTemplate Directory"):
-        shutil.copytree(Path("packagingTemplate") / i, Path("build") / i)
+        if not os.path.exists(Path("build") / i):
+            shutil.copytree(Path("packagingTemplate") / i, Path("build") / i)
 except Exception:
     log.exception(f"Error in copying packagingTemplate to build")
     exit(1)
@@ -59,10 +68,12 @@ except Exception:
     log.exception(f"Error while copying flags")
     exit(1)
 
-try:
-    shutil.rmtree(Path("build/__whl_cache__"))
-except Exception:
-    log.exception(f"Error while removing __whl_cache__")
+if not args.debug:
+    try:
+        shutil.rmtree(Path("build/__whl_cache__"))
+        log.debug(f"Removed __whl_cache__ folder")
+    except Exception:
+        log.exception(f"Error while removing __whl_cache__")
 
 log.info(f"Successfully Finished packaging")
